+++ Stripe/StripeiOSTests.xctestplan
@@ -24,7 +24,6 @@
   },
   "testTargets" : [
     {
-      "parallelizable" : true,
       "skippedTests" : [
         "PaymentMethodMessagingViewSnapshotTests",
         "STPCardBINMetadataTests",
